Updates in CSS/Readability/Slider
This is and update mainly for the CSS of the project migrating to the skeleton framework to make it more presentable.
The readbility of the main php file has been improved and minor changes to the slide function in the js have been made in order to align it to using the number of entrants as the area of the circle marker in contrast to using it as the radius.